Add support for promotion/coupon codes in checkout
This commit adds support for Stripe's newly-added promotional code parameter when creating a new Stripe Checkout session. I've defaulted this value to `true` because it's fairly unobtrusive.

The Stripe documentation on adding customer-facing promotion codes:
https://stripe.com/docs/payments/checkout/set-up-a-subscription#coupons

I've tested this with a local installation and found that it adds the promo code UI as expected (see included screenshots with PR). I then completed the checkout process with a coupon and found that it successfully applied the coupon - woohoo!

One thing worth noting: it doesn't appear that the coupon information shows up in the Members section, just the pricing (e.g. $50/yearly, but no mention if I use a 100% coupon). Not sure how to solve that one, but being able to view the subscription in Stripe from the link on the member page seems like an acceptable compromise!