Removed setting existing legacy prices to inactive
refs https://github.com/TryGhost/Members/commit/5ea54407e840a1c89b2211edc33bfddce026b32c

The current implementation of setting legacy prices as inactive has a bug where it also sets the active price to inactive if only one of monthly/yearly price is edited while the other is not changed. This is a temporary reversal till we can narrow down the issue and fix to correctly set legacy plans to inactive.