Updated DomainEvents to be non-singleton
refs https://github.com/TryGhost/Team/issues/1438

Use this package as a singleton runs into issues when multiple
packages are depending on it. If we have multiple versions of this
package in use by dependencies, those using different version will be
unable to communicate, rendering this package essentially useless.

We're making a breaking change here, rather than just updating Ghost
to pass down the instance, as that will help us find missed packages
easier, as renovate will let us know.

As this is a breaking change to the interface, this will be a major
version bump.